🐛 fix: skill_evolution inner attr + gateway auth borrow
- skill_evolution: #![allow] 改为顶部 crate 级 attr，修复 inner attr 编译错误
- gateway auth: or_insert_with 闭包不能同时 drop 和 borrow map，改用两步锁

+++ src-tauri/crates/gateway/src/auth.rs
@@ -201,7 +201,7 @@ impl KeyVerifyLimiter {
         // 先检查容量，必要时淘汰（在外层释放锁，闭包内不持有 map）
         {
             let mut map = self.failures.lock();
-            if map.len() >= self.max_failures {
+            if map.len() >= self.max_failures as usize {
                 self.evict_if_needed(&mut map);
             }
         }
